requirements.txt
elasticsearch==1.2.0
oauthlib>=0.7.2
requests>=2.4.3
requests-oauthlib>=0.4.1
six>=1.7.3
tweepy>=3.1.0
urllib3>=1.9.1
wsgiref>=0.1.2
pygithub>=1.25.2
